chore: rename npm scope @workspace/* → @workspace.sh/*
Renames all five workspace packages and every internal import,
dependency, doc reference, and demo script:

  @workspace/p2p-runtime       → @workspace.sh/p2p-runtime
  @workspace/ucan-boundary     → @workspace.sh/ucan-boundary
  @workspace/portable-bootstrap→ @workspace.sh/portable-bootstrap
  @workspace/workspace         → @workspace.sh/workspace
  @workspace/p2p-spike-node    → @workspace.sh/p2p-spike-node

Mirrors the workspace.sh domain. All packages are private, so the
dotted scope is fine for local workspace linking. NOTE: a dotted
scope can't be published to the public npm registry (org/scope names
are dotless) — if these ever go public the scope would become
@workspace-sh/* (hyphen). Trivial to flip, fully scripted.

Lockfile regenerated; 110 tests green across four packages; typecheck
clean (aside from the two pre-existing p2p-runtime errors unrelated to
this change). demo:acme verified under the new scope.
