+++ apps/node/src/demos/acme-org.ts
@@ -26,16 +26,16 @@ import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { createRequire } from 'node:module';
 
-import { NodeRuntime } from '@workspace/p2p-runtime/node';
-import { open, encryptedLog } from '@workspace/p2p-runtime';
-import { principalFromSeed } from '@workspace/ucan-boundary';
+import { NodeRuntime } from '@workspace.sh/p2p-runtime/node';
+import { open, encryptedLog } from '@workspace.sh/p2p-runtime';
+import { principalFromSeed } from '@workspace.sh/ucan-boundary';
 import {
   createBundle,
   consumeBundle,
   writeBundleFolder,
   readBundleFolder,
   type CapabilityDescriptor,
-} from '@workspace/portable-bootstrap';
+} from '@workspace.sh/portable-bootstrap';
 
 const require = createRequire(import.meta.url);
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
