+++ packages/p2p-runtime/src/index.ts
@@ -1,4 +1,4 @@
-// Public surface of @workspace/p2p-runtime.
+// Public surface of @workspace.sh/p2p-runtime.
 
 export { createRuntime } from './runtime.ts';
 export type {
@@ -12,7 +12,7 @@ export type {
 } from './types.ts';
 
 // Crypto primitives — exposed because they're independently useful (e.g. to
-// @workspace/portable-bootstrap which composes them into the bundle format).
+// @workspace.sh/portable-bootstrap which composes them into the bundle format).
 export { didFromSeed, didFromPublicKey, publicKeyFromDid, keyPairFromSeed } from './did.ts';
 export { wrap, unwrap, WRAP_OVERHEAD } from './wrap.ts';
 export { seal, open, SEAL_OVERHEAD, SEAL_KEY_BYTES } from './seal.ts';
