+++ packages/portable-bootstrap/tests/bundle.test.ts
@@ -1,4 +1,4 @@
-// Tests for @workspace/portable-bootstrap.
+// Tests for @workspace.sh/portable-bootstrap.
 //
 // End-to-end exercises: create a bundle for a small org, consume it as each
 // recipient, verify the keys round-trip. Plus tamper-detection tests so the
@@ -8,7 +8,7 @@ import { test } from 'node:test';
 import assert from 'node:assert/strict';
 import { createRequire } from 'node:module';
 
-import { principalFromSeed } from '@workspace/ucan-boundary';
+import { principalFromSeed } from '@workspace.sh/ucan-boundary';
 import {
   createBundle,
   consumeBundle,
